Initial commit to break up code into different modules
